requirements.txt
Django==4.2.13
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
python-decouple==3.8
pysqlite3==0.5.3
django-extensions==3.2.3
django-filter==24.2
django_constance==3.1.0
celery==5.4.0
redis==5.0.7
